+++ lib/faastruby/cli/commands/function/deploy_to.rb
@@ -13,9 +13,22 @@ module FaaStRuby
           load_credentials(exit_on_error: false)
         end
 
+        def ruby_runtime?
+          @yaml_config['runtime'].nil? || @yaml_config['runtime'].match(/^ruby/)
+        end
+
+        def crystal_runtime?
+          @yaml_config['runtime'].match(/^crystal/)
+        end
+
         def run
           create_or_use_workspace
-          FaaStRuby::CLI.error('Please fix the problems above and try again') unless bundle_install
+          if ruby_runtime?
+            FaaStRuby::CLI.error('Please fix the problems above and try again') unless bundle_install
+          end
+          if crystal_runtime?
+            FaaStRuby::CLI.error('Please fix the problems above and try again') unless shards_install
+          end
           tests_passed = run_tests
           FaaStRuby::CLI.error("Deploy aborted because tests failed and you have 'abort_deploy_when_tests_fail: true' in 'faastruby.yml'") unless tests_passed || !@abort_when_tests_fail
           puts "Warning: Ignoring failed tests because you have 'abort_deploy_when_tests_fail: false' in 'faastruby.yml'".yellow if !tests_passed && !@abort_when_tests_fail
@@ -57,6 +70,12 @@ module FaaStRuby
             sleep 2
             spinner.stop("Done!")
           end
+        end
+
+        def shards_install
+          puts '[build] Verifying dependencies'
+          return true unless File.file?('shard.yml')
+          system('shards check') || system('shards install')
         end
 
         def bundle_install