+++ lib/faastruby/server.rb
@@ -1,10 +1,12 @@
 module FaaStRuby
   PROJECT_ROOT = Dir.pwd
+  require 'faastruby/server/concurrency_controller'
   require 'faastruby/server/errors'
   require 'faastruby/server/event_channel'
   require 'faastruby/server/subscriber'
   require 'faastruby/server/event_hub'
   require 'faastruby/server/runner_methods'
+  require 'faastruby/server/function_object'
   require 'faastruby/server/runner'
   require 'faastruby/server/event'
   require 'faastruby/server/response'