+++ lib/faastruby/server/function_object.rb
@@ -1,0 +1,9 @@
+module FaaStRuby
+  class FunctionObject
+    include RunnerMethods
+    attr_reader :path
+    def initialize(path)
+      @path = path
+    end
+  end
+end