+++ lib/faastruby/spec_helper.rb
@@ -2,6 +2,7 @@ $LOAD_PATH << Dir.pwd
 module FaaStRuby
   require 'faastruby/server/errors'
   require 'faastruby/server/runner_methods'
+  require 'faastruby/server/function_object'
   require 'faastruby/server/event'
   require 'faastruby/server/response'
   ##########
@@ -28,5 +29,8 @@ module FaaStRuby
   module SpecHelper
     include FaaStRuby
     include RunnerMethods
+    def publish(channel, data: nil)
+      true
+    end
   end
 end